pyproject.toml
[tool.poetry]
name = "juniordeepsequence"
version = "0.1.0"
description = ""
authors = ["walle <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
yahoo-fin = "^0.8.6"
pandas = "^1.1.4"
lxml = "^4.6.1"
requests-html = "^0.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
